LibELF: Remove the unused Image::Section::relocations function
This function assumes that all sections have a relocation section
associated with them. This is not the case in most executables.
It also assumes that the section is called ".rel<section name>". This is
also not something we should rely on. And RELA sections usually start
with ".rela", not ".rel".