Kernel: Create coredump by using the crashed process VFS root context
Resolve a regression caused by 01e1af732ba2eb7dd0c31d71fe9b008d8f5e9d42.

This unbreaks coredump generation, because we need to use the VFS root
context of the crashed process and not of the FinalizerTask, as it will
hold an empty VFS root context that is assigned to kernel processes.